Squish actualize.
Squished the `actualize` function first by changing the day format so
that the algorithm for finding the first occurance of particular day of
the week after a certain date was the same as finding the last occurance
of a particular day of the week in the month.

The day of the week is converted to an integer, with `0` for Sunday in
our `bin/zones.js` preprocessor. This eleminates the day of week name
constants from `lib/timezone.js`.

We use an offset to determine the direction of the search. When
searching forward, we start from the date that the day of week must be
greater than or equal to. When starting backwards we start with a date
of `-31`. We multiply that by our increment. The JavaScript `Date`
object will place us in the subsequent month. We move backward until
we're the correct month. As we increase our date it approaches zero and
when we multiply by our increment

We stop on the day of week
offset will decrement from t

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          11509/11.24  5448/05.32  3354/03.28  2415/02.36
    zones/America/Detroit.js  2987/02.92  2476/02.42   556/00.54   543/00.53
    zones/Europe/Amsterdam.js 5709/05.58  4755/04.64   766/00.75   743/00.73