Convert to plain text `README`.
Replace our jumbled, confused and out of date markdown `README.md` with
a plain ASCII `README` that directs the visitor to the GitHub:Pages.

Fixed bug in `synopsis.js`.

This is an attempt to keep messages between here and issues up to date
by closing this with the command below. I'm curious to see how much this
alters the git history.

Closes #37.