Size savings searching clock transition day.
In `actualize`, when searching for a last occurrence of a day of the
week in a particular month, it is not necessary to test that the month
is indeed the desired month. This test was because we always set the
last day of the month to 31. We now set it to the actual last day of the
month for the specific month.

Fortunately, there is no rule that requests a clock transition on a last
day of the month in February, where the day of the week happens to also
be the same day of the week March 1st on a non-leap year. We assert that
this is true in `bin/zones.js`. If it were the case, then removing the
month test would cause the lookup to fail because on a non-leap year,
creating a day with a month of February and a date of 29 would create a
March 1st date, the day of the week would be correct, and the loop would
exit on the 1st March, instead of February 22nd.

Also in `actualize`, we indicate an exact date by setting the desired
day of the week to seven because less than seven compresses better than
greater than or equal to zero.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          15533/15.17  7904/07.72  4176/04.08  3019/02.95