Minify main arguments loop.
In `convert`, the loop through the arguments reuses `i` to track the
current index instead of `index` reducing the number of variables,
making the library slightly smaller when minified and gzipped.

Also in `convert`, the last `else if` in the `if` `else` later has been
turned into an `else`. The types have been exhausted at this point, it
is obviously an object, no test is necessary.

Earlier, I'd removed a test to see if the locale was null. The test is
no longer necessary because of the merge with `date.js`. It used to be
the case that `date.js` would introduce locales, so no locale would be
defined until `date.js` was included. Now `en_US` is the default locale.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          15456/15.09  7880/07.70  4146/04.05  3001/02.93