Minify date type detection.
Moved removed the test for number in the `if` `else` latter and made the
conversion to `number` the `else`. If the date is not already a number,
and it is an object of any sort, the `valueOf` method of the object will
be called. This will convert JavaScript `Date` objects, so we can accept
`Date` objects at a super low cost in size.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          15482/15.12  7859/07.67  4162/04.06  2996/02.93