Correctly format zone offset padding.
Zone offsets are colon delimited and have the oddball `:` flag that
exists only for the zone offsets. Zone offsets can have negative values,
so they are always formatted with a sign. The padding flags still apply
to the zone offsets, but their logic is different to accommodate the
colons and sign.

All the other functions can share padding logic, so their padding can be
applied after they return their integer value. Because it's a special
case, the handling is placed in the `z` format function itself.

More logic means more bytes that are hard to squish. File size has
jumped. When I return to remove `%$` and replace it `^z`, I hope to lose
some of this girth.

At least the price we pay for `^z` will not seem so dear. It is our
customization, for which we'd have to make the `z` format function
bigger and more complicated. Now we have the `flag` property in the `z`
format specifier to support our de facto standard, so we're only adding
formatting logic.

Closes #39.
Closes #40.