Format specifier descriptions.
Created a listing of the format specifiers with descriptions of what
they emit. Tedious job that I've been putting off. Considered just
linking to GNU `date` documentation, but hosting this table as part of
Timezone documentation will be good for Timezone.

Descriptions are my own. Considering breaking them out into a table with
padding and length specified as part of the table, so that the tedious
phrases are not repeated &mdash; i.e. "two digits, zero padded".

Formatting comes later. Happy to have the writing done.