Build on OS X.
Now building on OS X.

Unable to compile `zic` from source on OS X, so I made `zic` a separate make
target. If the `Makefile` cannot find a zic compiled from source, it will use
the system `zic`.

The `Makefile` will use either `md5` or `md5sum` to verify the downloaded
IANA Timezone Database source.

On OS X, `npm` could not install package depdendencies because `depdendencies`
was missing from the `package.json`. It's empty, we have only `devDependencies`,
but NPM was counting on it to be there.

Updated the `README` with a **HACKING** section.

Closes #31.