UTC with partial functions, TODOs removed.
If you'd created a Timezone function with a different default timezone
it was impossible to override the default timezone with UTC. The default
time zone entry for the default context is a UTC time zone entry, but
their is no UTC time zone rule set in the default context. I added a
value of `true` to the `UTC` property of the default context so that the
argument processing branch would accept UTC as a correct timezone. The
conversion functions short circuit on UTC and return the value they've
been given, so they do not need a rule set.

Remove TODOs and created issues where necessary.

Some TODOs where complaining about the choice of format specifier in
unit tests. I satisified some and deleted others.

Closes #51.
Closes #46.