Fixed length date array format.
Instead of looking for a `'+'` or `'-'` character in the date array, if
the eighth position in the array is non-zero, it is assumed to be the
offset direction expressed as either `-1` for locations West of
Greenwich `1` for locations East of Greenwich.

Any array element left undefined or otherwise untruthy will be set to
`0`, so it is still possible to do a shorthand date array with, for
example, only the year, month and date.

Removed the logic that would search for `'+'` and `'-'` and interpret
the findings.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              14941/14.59  7598/07.42  4086/03.99  2922/02.85

Closes #52.