Complete RFC 2822 date parser.
Full implementation of an RFC 2822 date parser as a function that
returns a **Timezone** date array. Most of what needs to be said about
the implementation is said in the Docco of `rfc822.js`.

I'll move the test into place in a subsequent commit. I want to make
sure that `git` notices the rename.

    File                             Raw         Min         Zip     Min/Zip
    rfc822.js                 7638/07.46  3187/03.11  3347/03.27  1613/01.58