Return partial if date fails.
Was feeling that it was important to go right ahead and ignore a date
that does not parse correctly, or an array date that resolves to `NaN`,
but as I commit the change, I'm having a change of heart.

Why not make a best effort to validate the users data? If they provide a
string that begins with a number, then it must be a date. If it doesn't
parse, then we ought to throw an exception. It is not a recoverable
exception, but it will aid in debugging.

I'm retreating because I need to find a way to report an RFC 3999 date
string that given a time zone detects that the time does not exist in
the time zone. Return NaN? Well, what if there is a date format?

An exception is the only reasonable response. Perhaps returning NaN if
there is no format, but raising an exception if there is one?

In any case, now that date array is unambiguous, the intent of the
caller is clear. It makes more sense to raise an exception if the
contents are not numeric, than to return a partial.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              14855/14.51  7679/07.50  4074/03.98  2954/02.88