Unambiguous zone rule sets.
Previously, the test to see if a member of the context was a zone rule
set was to test if the member was an array. Zones are arrays, but so are
the daylight savings time rules. To resolve this ambiguity, the zone
rule set arrays now have the string `"z"` as their first element.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              14855/14.51  7675/07.50  4078/03.98  2959/02.89