Reorganize `actualize` for savings.
Using a single variable to work through `actualize`. It is used first to
find the actualized date, then to hold the actualized structure.

Some reworking of the setting of the unspecified clock resulted in some
byte savings, including simply testing if the `posix` property is set,
instead of testing the `clock` name property. This will result in an
error if we actualize a date to New Year's millisecond, 1970. All tests
pass, so unless new research turns up a previously unknown history clock
transition, we should be okay.

Renaming the actualized object member `sortable` to `sort` brings byte
savings when it smooshes with `Array.sort`.

Note that there is an unused variable here. If I remove it I gain a
byte, so I'm going to leave it in. Silly, I know.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              14613/14.27  7470/07.29  4002/03.91  2882/02.81

Closes #63.