Browser resources, docco update.
Created `data.js` to publish Timezone locale and time zone data as AMD
friendly modules, JSONP with a fixed callback name, and JSON, which
isn't much use for a browser application, but might be useful on other
platforms.

I'd like to get particular about file extensions, adding `.json` to the
JSON files, leaving the AMD and JSONP files `.js`. Depends on what
GitHub:Pages sends for a mine-type when it serves `.json`.

Created a test server in `server.js`, and a script for when I forget how
to run in in `server.sh`.

Create a small AMD based program in `amd.html` that reads the resources
in `scripts` to test AMD. Possible start of a fancier GitHub:Pages
landing page.

Update the `synopsis.js` based documentation.

Closes #23.