Deleted `notes` directory.
Created a `notes` branch and pushed it to GitHub prior to deleting the
`notes` directory.

The `notes` directory is unsightly in the GitHub directory listing. It
has some crufty writing, an early version of `synopsis.js`, a program
written in Java to test the date formatting libraries of that platform,
and a well written glossary of the format specifiers.

Ideally the writing would get integrated into the documentation and the
Java program could live alone on an detached branch for posterity.