Renamed `zones` intermediate directory to `build`.
Although the `zones` intermediate directory will only ever contain
intermediate files pertaining to timezone data, it is more important to
convey the temporary nature of the directory, more important than
conveying the contents. In fact, you want to play down the contents to
stress the temporary nature of it all.