Reintroduce `.npmignore`.
My new attitude toward `.AppleDouble` directories is to accept their
omnipresence and to ignore them. Instead of having to remember that I
need to delete the little devils before publishing to NPM, I'll use an
`.npmignore`.

Seems a bother to have an `.npmignore` for a module directory that is
the output of a build process. There are no source files to ignore, only
these meta files that exist for the sake of Netatalk. Yet trying to
delete them is whack-a-mole; an omnipresent race condition.

The ability to use my OS X programs on my Linux and Illumos virtual
machines is a boon to my productivity. If I'm going to pay the price for
using Netatalk, I'm going to pay it in full. Could it be that accepting
that hidden directories are always in the way and dealing with them is a
passage of sorts?