Fix minification report. Use default Ruby.
Fixed the minification report on Travis CI to account for the relocation
of the `timezone.js` source file.

The default Ruby at Travis CI is now 1.9.3p286 which supports the `%::z`
format specifier, so we no longer need to install a Ruby with `apt-get`
from a Personal Package Archive.

Earlier on this branch of development, at commit 93d3cb2, we updated the
`Makefile` to use multiple targets to good effect. That closes #130.

With this update to `util/test` we will close #135.

Closes #135.
Closes #130.