Include links width slashes in their names.
Rather than store a link in the JavaScript zone information file, I
instead create a full JavaScript zone information file. When someone is
loading the entire database at once, this is going to be wasteful, but
when someone is requesting timezones over the web as needed, this is
going to be most efficient.

Because I'm using one object to preserve the state of the `tz` function,
I've counted on the fact that the active set of normalized timezone
names all have a slash in them, while all of the zone rules do not.
When adding timezones links, the created a collision when the legacy
timezone name `Cuba` collided with the DST rules named `Cuba`.

The obvious solution to this is to store the names with a `z:` or `r:`
prefix in the enclosed timezone object.

The `US/Pacific` timezone mentioned in #144 is now visible.