Use `require` for resource indexes.
Use `require` for resource indexes. An index is used for locales and for
each of the timezone groupings such as "America" or "America/Indiana".
The index will include all of the locales or timezones in the directory
for the grouping. It would use the `fs.readdirSync` to get a directory
listing and loop through it invoking `require`. This is foolish. It
defeats static analysis and therefore defeats tools like Browserify.

I've updated the `Makefile` to build an `index.js` that performs a
require using string literals.

Closes #259.