Fix 2016g parse errors.
Removing an extra run of the skip list creation function fixes a problem
with Europe/Moscow. Probably should make a defensive copy of the data at
each invocation, but want to check this in to show the simple change
that fixes the build.

Also, there is a change to Asia/Dushanbe that causes us to neglect to
parse the abbreviation format. It is not parsed when the previous entry
had no rules to apply. When that is the case we go and split the format
if it needs splitting and apply the correct abbreviation based on
whether savings are present at the start of the period.

Closes #278.