Simplify after date transition boundaries.
When we are looking for the first day of the week greater than a certain
date, we already have the start date. For some reason I was starting at
the first of the month instead of just starting at the given date. I was
performing the greater than or equal too in the loop, instead of just
looking for the day of the week. This will throw off the seek if the
first day of the week (i.e. first Friday on or after the 29th) occurs in
the next month.