First day of week less than or equal to, 2019b.
2019b uses a concept of first day of week less than or equal to a
certian date that had not yet been encountered. Supporting it required
some slight adjustment to `timezone.js` itself. Working backward to a
last day of the week from the end of the month had already been
implemented, but it was implemented under the assumption that the search
would not continue into the previous month. Now we're starting our
search at the first of the month so we must accommodate the wrap.

This was necessary to support IANA TZ 2019b, also included in this
commit.

Closes #328.
Closes #327.